Solving infinite loop. Solving wrong counter behaviour on ft_print_char. Solving wrong behaviour of passing of va_list args not as pointer.

+++ srcs/ft_print_pointer.c
@@ -6,7 +6,7 @@
 /*   By: fyudris <fyudris@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/17 19:44:41 by fyudris           #+#    #+#             */
-/*   Updated: 2024/12/17 19:51:09 by fyudris          ###   ########.fr       */
+/*   Updated: 2024/12/18 14:08:13 by fyudris          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,15 +26,15 @@ int ft_print_pointer(t_format *f, void *ptr)
 	if (!hex_str)
 		return (-1);
 	// Handle right padding
-	if (!f->minus && f->width > (ft_strlen(hex_str) + 2))
-		count += ft_print_padding(f->width - (ft_strlen(hex_str) + 2), ' ');
+	if (!f->minus && f->width > ((int)ft_strlen(hex_str) + 2))
+		count += ft_print_padding(f->width - ((int)ft_strlen(hex_str) + 2), ' ');
 	// Print prefix
 	count += ft_putstr_fd("0x", 1);
 	// Print the hex string
 	count += ft_putstr_fd(hex_str, 1);
 	// Handle left padding
-	if (f->minus && f->width > (ft_strlen(hex_str) + 2))
-		count += ft_print_padding(f->width - (ft_strlen(hex_str) + 2), ' ');
+	if (f->minus && f->width > ((int)ft_strlen(hex_str) + 2))
+		count += ft_print_padding(f->width - ((int)ft_strlen(hex_str) + 2), ' ');
 	free(hex_str);
 	return (count);
 }
